// src/shared/types/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number;
  title: string;
  slug: string;
  layout?:
    | (
        | {
            layout?: ('imageRight' | 'imageLeft') | null;
            title: string;
            description?: string | null;
            image: number | Media;
            extraContent?:
              | (
                  | {
                      items?:
                        | {
                            item?: string | null;
                            id?: string | null;
                          }[]
                        | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'list';
                    }
                  | {
                      title?: string | null;
                      paragrahe?: string | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'card';
                    }
                  | {
                      label?: string | null;
                      link?: string | null;
                      variant?: ('primary' | 'secondary' | 'ternary') | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'button';
                    }
                  | {
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'inputEmail';
                    }
                )[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'mediaAndText';
          }
        | {
            id?: string | null;
            blockName?: string | null;
            blockType: 'supportBlock';
          }
        | {
            title: string;
            plans?:
              | {
                  title: string;
                  price: string;
                  description?: string | null;
                  items?:
                    | {
                        item?: string | null;
                        id?: string | null;
                      }[]
                    | null;
                  button: {
                    label?: string | null;
                    link?: string | null;
                  };
                  id?: string | null;
                }[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'pricingPlansBlock';
          }
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
}